LayoutTests/fast/mediastream/RTCPeerConnection-stats-expected.txt

Tests the RTCPeerConnection stats interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new webkitRTCPeerConnection(null) did not throw exception.
PASS pc.getStats(statsHandler1) did not throw exception.
PASS statsHandler1 was called
PASS status is non-null.
PASS result.length is 0
PASS getUserMedia({audio:true, video:true}, gotStream) did not throw exception.
PASS Got a stream.
PASS pc.getStats(statsHandler2) did not throw exception.
PASS statsHandler2 was called
PASS result.length is >= 2
PASS local.length is 1
PASS timestamp is >= startTime
PASS local[0].stat("type") is "audio"
PASS successfullyParsed is true

TEST COMPLETE
